refactor: rename internal packages parser→schema, scanner→refs
Rename internal/parser → internal/schema and internal/scanner → internal/refs
to better reflect their purpose (schema definition extraction vs reference
detection), rather than generic programming terms.

File renames within each package follow the same principle — organized by ORM
first, then named by what they extract/detect rather than by input file name:

internal/schema/:
  schema_rb.go      → rails_schema.go  (parses db/schema.rb)
  migrations.go     → rails_migrations.go  (replays db/migrate/)
  models.go         → django.go  (parses models.py)

internal/refs/:
  scanner.go splits into:
    scan.go     — shared infra (scanFiles, dedupeByLine, lineAt, walkFn)
    rails.go    — Ruby reference detection
    django.go   — Python/Django reference detection

Closes #124
