+++ internal/refs/rails.go
@@ -1,0 +1,146 @@
+package refs
+
+import (
+	"strings"
+
+	sitter "github.com/smacker/go-tree-sitter"
+	"github.com/smacker/go-tree-sitter/ruby"
+
+	"github.com/shinagawa-web/colref/internal/orm"
+)
+
+// RubySkipDirs is the set of directory names skipped when scanning Ruby projects.
+var RubySkipDirs = map[string]bool{
+	"spec":         true,
+	"test":         true,
+	"vendor":       true,
+	"migrate":      true,
+	"node_modules": true,
+}
+
+// RubyScanner implements orm.ReferenceScanner for Ruby codebases.
+type RubyScanner struct{}
+
+// Scan implements orm.ReferenceScanner.
+func (RubyScanner) Scan(dir, fieldName string) ([]orm.Reference, int, error) {
+	return ScanRuby(dir, fieldName)
+}
+
+// SkipDirs implements orm.ReferenceScanner, returning a defensive copy.
+func (RubyScanner) SkipDirs() map[string]bool {
+	copy := make(map[string]bool, len(RubySkipDirs))
+	for k, v := range RubySkipDirs {
+		copy[k] = v
+	}
+	return copy
+}
+
+// ScanRuby walks dir and returns every method-call node whose method name
+// equals fieldName, along with the total number of .rb and .erb files examined.
+func ScanRuby(dir, fieldName string) ([]Reference, int, error) {
+	return scanFiles(dir, fieldName, map[string]func([]byte) []byte{
+		".rb":  nil,
+		".erb": erbToRuby,
+	}, ruby.GetLanguage(), walkNodeRuby, RubySkipDirs)
+}
+
+// walkNodeRuby matches Ruby method calls (call nodes) where the method name
+// equals fieldName, e.g. user.email → matches "email".
+// A receiver is required to avoid false positives on standalone method calls
+// like raw(string) or send(msg) that share a name with the target field.
+func walkNodeRuby(node *sitter.Node, src []byte, lines [][]byte, fieldName, file string, refs *[]Reference) {
+	if node.Type() == "call" {
+		method := node.ChildByFieldName("method")
+		receiver := node.ChildByFieldName("receiver")
+		if method != nil && receiver != nil && method.Content(src) == fieldName {
+			methodRow := int(method.StartPoint().Row)
+			text := node.Content(src)
+			if int(node.StartPoint().Row) != methodRow {
+				text = strings.TrimSpace(lineAt(lines, methodRow))
+			}
+			*refs = append(*refs, Reference{
+				File: file,
+				Line: methodRow + 1,
+				Text: text,
+			})
+		}
+	}
+	for i := 0; i < int(node.ChildCount()); i++ {
+		walkNodeRuby(node.Child(i), src, lines, fieldName, file, refs)
+	}
+}
+
+func erbToRuby(src []byte) []byte {
+	out := make([]byte, len(src))
+	inRuby := false
+	inComment := false
+	inString := byte(0) // '"' or '\'' when inside a Ruby string literal
+	escaped := false
+	i := 0
+	for i < len(src) {
+		if inComment {
+			if i+1 < len(src) && src[i] == '%' && src[i+1] == '>' {
+				out[i], out[i+1] = ' ', ' '
+				i += 2
+				inComment = false
+			} else if src[i] == '\n' {
+				out[i] = '\n'
+				i++
+			} else {
+				out[i] = ' '
+				i++
+			}
+		} else if inRuby {
+			if inString != 0 {
+				out[i] = src[i]
+				if escaped {
+					escaped = false
+				} else if src[i] == '\\' {
+					escaped = true
+				} else if src[i] == inString {
+					inString = 0
+				}
+				i++
+			} else if i+1 < len(src) && src[i] == '%' && src[i+1] == '>' {
+				// Terminate the ERB block as a Ruby statement so two
+				// adjacent <%= a %> <%= b %> tags on the same source line
+				// don't collapse into a single ambiguous call when the
+				// surrounding HTML is converted to whitespace. Without the
+				// `;`, tree-sitter parses the second tag as a continuation
+				// of the first, dropping its `call` nodes (issue #64).
+				out[i], out[i+1] = ';', ' '
+				i += 2
+				inRuby = false
+			} else {
+				if src[i] == '"' || src[i] == '\'' {
+					inString = src[i]
+				}
+				out[i] = src[i]
+				i++
+			}
+		} else {
+			if i+1 < len(src) && src[i] == '<' && src[i+1] == '%' {
+				out[i], out[i+1] = ' ', ' '
+				i += 2
+				if i < len(src) && src[i] == '#' {
+					out[i] = ' '
+					i++
+					inComment = true
+				} else {
+					if i < len(src) && (src[i] == '=' || src[i] == '-') {
+						out[i] = ' '
+						i++
+					}
+					inRuby = true
+				}
+			} else if src[i] == '\n' {
+				out[i] = '\n'
+				i++
+			} else {
+				out[i] = ' '
+				i++
+			}
+		}
+	}
+	return out
+}
