+++ internal/schema/rails_migrations.go
@@ -1,4 +1,4 @@
-package parser
+package schema
 
 import (
 	"context"
