+++ internal/schema/rails_migrations_test.go
@@ -1,4 +1,4 @@
-package parser
+package schema
 
 import (
 	"context"
