+++ internal/schema/rails_schema.go
@@ -1,4 +1,4 @@
-package parser
+package schema
 
 import (
 	"context"
