+++ internal/schema/rails_schema_test.go
@@ -1,4 +1,4 @@
-package parser
+package schema
 
 import (
 	"context"
